Simplify authenticated dispersal
